+++ app/src/main/java/com/app/arcabyolimpo/data/remote/api/ArcaApi.kt
@@ -14,6 +14,7 @@ import com.app.arcabyolimpo.data.remote.dto.password.UpdatePasswordResponseDto
 import com.app.arcabyolimpo.data.remote.dto.password.VerifyTokenResponseDto
 import com.app.arcabyolimpo.data.remote.dto.productbatches.ProductBatchDto
 import com.app.arcabyolimpo.data.remote.dto.product.ProductDto
+import com.app.arcabyolimpo.data.remote.dto.product.ProductRegisterInfoDto
 import com.app.arcabyolimpo.data.remote.dto.productbatches.ProductBatchModifyDto
 import com.app.arcabyolimpo.data.remote.dto.productbatches.ProductBatchRegisterDto
 import com.app.arcabyolimpo.data.remote.dto.supplies.AcquisitionDto
@@ -35,6 +36,7 @@ import com.app.arcabyolimpo.data.remote.dto.workshops.DeleteResponseWorkshopDto
 import com.app.arcabyolimpo.data.remote.dto.workshops.DeleteWorkshopDto
 import com.app.arcabyolimpo.data.remote.dto.workshops.WorkshopDto
 import com.app.arcabyolimpo.data.remote.dto.workshops.WorkshopsListDto
+import com.app.arcabyolimpo.data.remote.dto.workshops.WorkshopResponseDto
 import okhttp3.MultipartBody
 import okhttp3.RequestBody
 import retrofit2.Response
@@ -172,7 +174,7 @@ interface ArcaApi {
     @GET("workshop/{id}")
     suspend fun getWorkshop(
         @Path("id") id: String,
-    ): WorkshopDto
+    ): WorkshopResponseDto
 
     @POST("workshop/add")
     suspend fun addWorkshop(
@@ -261,6 +263,11 @@ interface ArcaApi {
         @Body batch: ProductBatchModifyDto,
     )
 
+    @DELETE("productBatch/{id}")
+    suspend fun deleteProductBatch(
+        @Path("id") id: String,
+    )
+
     // Products --------------------------
 
     @Multipart
@@ -287,4 +294,23 @@ interface ArcaApi {
         @Query("q") query: String,
     ): List<ProductDto>
 
+    @GET("product/add")
+    suspend fun getProductFilters(): ProductRegisterInfoDto
+
+    @GET("product/disponible")
+    suspend fun getProductsByAvailability(
+        @Query("disponible") disponible: Int,
+    ): List<ProductDto>
+
+
+    @GET("product/workshop")
+    suspend fun getProductsByWorkshop(
+        @Query("idTaller") idTaller: String,
+    ): List<ProductDto>
+
+    @GET("product/order")
+    suspend fun getProductsOrderedByPrice(
+        @Query("orderBy") orderBy: String,
+        @Query("direction") direction: String,
+    ): List<ProductDto>
 }